Add files via upload
Now has 'Float' datatype instead of 'Integer' for the 'price' column in the database

+++ Springboot CRUD Project/demo/src/main/resources/db/migration/V1__create_products_table.sql
@@ -2,6 +2,6 @@ CREATE TABLE if not exists products(
     id int PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     name varchar(255),
     description varchar(255),
-    price Integer,
+    price Float,
     product_type varchar(255)
 );
